app/requirements.txt
cassandra-driver==3.29.1
click==8.1.7
geomet==0.2.1.post1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
wheel==0.43.0
